pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "cairosvg>=2.8.2",
    "matplotlib>=3.10.3",
    "pillow>=11.2.1",
]
